All routes implement JWT. Updated Models to link user to their posts. Fixed Login route.

+++ controllers/contacts.controller.js
@@ -3,7 +3,8 @@ const Response = require('../Utils/ResponseBuilder');
 
 const getContacts = async (req, res, next) => {
     try {
-        const contacts = await Contact.find({}).limit(10);
+        const { _id } = req.user;
+        const contacts = await Contact.find({ userId: _id }).limit(10);
         const response = Response.success("limit of 10 contacts", contacts);
         res.status(200).json(response);
     } catch (error) {
@@ -15,9 +16,10 @@ const getContacts = async (req, res, next) => {
 
 const getContact = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
-        const contact = await Contact.findById({ _id: id });
-        const response = Response.success(`Contact with id ${id} found`, contact);
+        const contact = await Contact.findById({ _id: id, userId: _id });
+        const response = Response.success(`Contact with id ${id} found`, [contact]);
         res.status(200).json(response);
     } catch (error) {
         const response = Response.failed(error, 404);
@@ -27,10 +29,17 @@ const getContact = async (req, res, next) => {
 
 const addContact = async (req, res, next) => {
     try {
-        const contactAdded = new Websites({ ...req.body });
-        const contact = await contactAdded.save();
-        const response = Response.success("Contact added", contact);
-        res.status(200).json(response);
+        const { _id } = req.user;
+        if (_id) {
+            const contactAdded = new Websites({ userId: _id, ...req.body });
+            const contact = await contactAdded.save();
+            const response = Response.success("Contact added", contact);
+            res.status(200).json(response);
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
+
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
@@ -39,11 +48,18 @@ const addContact = async (req, res, next) => {
 
 const updateContact = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
         const body = req.body;
-        const upd8tedContact = await Contact.findByIdAndUpdate({ _id: id }, { $set: body });
-        const response = Response.success(`Contact with ${id} was successfully updated`, upd8tedContact);
-        res.status(200).json(response);
+        if (_id) {
+            const upd8tedContact = await Contact.findByIdAndUpdate({ _id: id, userId: _id }, { $set: body });
+            const updated = await Contact.findById({ _id: id, userId: _id })
+            const response = Response.success(`Website with id ${id} updated`, [updated]);
+            res.status(200).json(response);
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
@@ -52,9 +68,15 @@ const updateContact = async (req, res, next) => {
 
 const removeContact = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
-        const data = await Contact.findByIdAndDelete({ _id: id });
-        res.status(204).end();
+        if (_id) {
+            const data = await Contact.findByIdAndDelete({ _id: id, userId: _id });
+            res.status(204).end();
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
