+++ controllers/notes.controller.js
@@ -3,7 +3,8 @@ const Response = require('../Utils/ResponseBuilder');
 
 const getNotes = async (req, res, next) => {
     try {
-        const notes = await Note.find({}).limit(10);
+        const { _id } = req.user;
+        const notes = await Note.find({ userId: _id }).limit(10);
         const response = Response.success("limit of 10 notes", notes);
         res.status(200).json(response);
     } catch (error) {
@@ -14,8 +15,9 @@ const getNotes = async (req, res, next) => {
 
 const getNote = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
-        const note = await Note.findById({ _id: id });
+        const note = await Note.findById({ _id: id, userId: _id });
         const response = Response.success(`Success, note with id ${id} found`, [note]);
         res.status(200).json(response);
     } catch (error) {
@@ -26,11 +28,17 @@ const getNote = async (req, res, next) => {
 
 const addNote = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const { text } = req.body;
-        const noteAdded = new Note({ text });
-        const note = await noteAdded.save();
-        const response = Response.success("Note added", [note]);
-        res.status(200).json(response);
+        if (_id) {
+            const noteAdded = new Note({ text, userId: _id });
+            const note = await noteAdded.save();
+            const response = Response.success("Note added", [note]);
+            res.status(200).json(response);
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
@@ -39,11 +47,18 @@ const addNote = async (req, res, next) => {
 
 const updateNote = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
         const body = req.body;
-        const upd8tedNote = await Note.findByIdAndUpdate({ _id: id }, { $set: body });
-        const response = Response.success(`Note with id ${id} updated`, [upd8tedNote]);
-        res.status(200).json(response);
+        if (_id) {
+            const upd8tedNote = await Note.findByIdAndUpdate({ _id: id, userId: _id }, { $set: body });
+            const updated = await Note.findById({ _id: id })
+            const response = Response.success(`Note with id ${id} updated`, [updated]);
+            res.status(200).json(response);
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
@@ -52,9 +67,15 @@ const updateNote = async (req, res, next) => {
 
 const removeNote = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
-        const data = await Note.findByIdAndDelete({ _id: id });
-        res.status(204).end();
+        if (_id) {
+            const data = await Note.findByIdAndDelete({ _id: id, userId: _id });
+            res.status(204).end();
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
