+++ controllers/websites.controller.js
@@ -3,7 +3,8 @@ const Response = require('../Utils/ResponseBuilder');
 
 const getWebsites = async (req, res, next) => {
     try {
-        const websites = await Websites.find({}).limit(10);
+        const { _id } = req.user;
+        const websites = await Websites.find({ userId: _id }).limit(10);
         const response = Response.success("limit of 10 websites", websites);
         res.status(200).json(response);
     } catch (error) {
@@ -14,8 +15,9 @@ const getWebsites = async (req, res, next) => {
 
 const getWebsite = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
-        const website = await Websites.findById({ _id: id });
+        const website = await Websites.findById({ _id: id, userId: _id });
         const response = Response.success(`Success, website with id ${id} found`, [website]);
         res.status(200).json(response);
     } catch (error) {
@@ -26,10 +28,17 @@ const getWebsite = async (req, res, next) => {
 
 const addWebsite = async (req, res, next) => {
     try {
-        const websiteAdded = new Websites({ ...req.body });
-        const website = await websiteAdded.save();
-        const response = Response.success("Website added", [website]);
-        res.status(200).json(response);
+        const { _id } = req.user;
+        if (_id) {
+            const websiteAdded = new Websites({ userId: _id, ...req.body });
+            const website = await websiteAdded.save();
+            const response = Response.success("Website added", [website]);
+            res.status(200).json(response);
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
+
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
@@ -38,11 +47,18 @@ const addWebsite = async (req, res, next) => {
 
 const updateWebsite = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
         const body = req.body;
-        const upd8tedWebsite = await Websites.findByIdAndUpdate({ _id: id }, { $set: body });
-        const response = Response.success(`Website with id ${id} updated`, [upd8tedWebsite]);
-        res.status(200).json(response);
+        if (_id) {
+            const upd8tedWebsite = await Websites.findByIdAndUpdate({ _id: id, userId: _id }, { $set: body });
+            const updated = await Websites.findById({ _id: id, userId: _id })
+            const response = Response.success(`Website with id ${id} updated`, [updated]);
+            res.status(200).json(response);
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
@@ -51,9 +67,15 @@ const updateWebsite = async (req, res, next) => {
 
 const removeWebsite = async (req, res, next) => {
     try {
+        const { _id } = req.user;
         const id = req.params.id;
-        const data = await Websites.findByIdAndDelete({ _id: id });
-        res.status(204).end();
+        if (_id) {
+            const data = await Websites.findByIdAndDelete({ _id: id, userId: _id });
+            res.status(204).end();
+        } else {
+            const response = Response.failed(null, 406, "Failed");
+            res.status(406).json(response);
+        }
     } catch (error) {
         const response = Response.failed(error, 406);
         res.status(406).json(response);
